test(e2e): extract terminal WS mock helper; diagnose desktop-selection
- Add e2e/helpers/mock-backends.ts (mockTerminalWebSocket) and use it from
  terminal-reconnect (DRY; validated still green on the Jetson prod build).
- desktop-selection: locator scoping done, WS mock wired, but re-fixme'd with a
  full diagnosis — the (40,40) right-click doesn't open the desktop context
  menu on the prod build (hit-testing/layout), which must be solved first.
  Tracked in #114.
